create test files in constructor, remove unnecessary tests, add aspect and slope test

+++ test.py
@@ -4,34 +4,40 @@ import windowagg.dem as dem
 from windowagg.agg_ops import Agg_ops
 import windowagg.aggregation as aggregation
 from image_generator import ImageGenerator
+import windowagg.config as config
 
 import unittest
 import math
 import os
 import shutil
+import random
 
 import numpy as np
 import rasterio
 
 class TestSlidingWindow(unittest.TestCase):
 
-    # used in tearDownClass method
-    test_dir = 'test_img/'
-    IMAGE_SIZE = 1024
-    NUM_AGGRE = 4
+    @classmethod
+    def setUpClass(self):
+        self.test_dir = 'test_img/'
+        self.image_size = 512
+        self.num_aggre = 4
+        self.num_test_points = 20
+        self.dtype = config.work_dtype
 
-    img_gen = ImageGenerator()
-    img_gen.test_dir = test_dir
+        self.img_gen = ImageGenerator(self.test_dir)
+        self.gauss = self.img_gen.gauss(self.image_size)
+        self.random = self.img_gen.random(self.image_size)
 
-    # remove test folder after testing completes
+
     @classmethod
     def tearDownClass(self):
-        shutil.rmtree(self.test_dir)
+        if (os.path.exists(self.test_dir)):
+            shutil.rmtree(self.test_dir)
 
     def test_aggregation(self):
-        path = self.img_gen.random()
-        with rasterio.open(path) as img:
-            arr = img.read(1).astype(float)
+        with rasterio.open(self.random) as img:
+            arr = img.read(1).astype(self.dtype)
 
         sum_all = aggregation.aggregate(arr, Agg_ops.add_all, 5, 0)
         partial_sum_all = aggregation.aggregate(arr, Agg_ops.add_all, 2, 0)
@@ -88,36 +94,30 @@ class TestSlidingWindow(unittest.TestCase):
         self.assertTrue(np.array_equal(partial_minimum, brute_minimum))
 
     def test_regression(self):
-        path = self.img_gen.random()
-        with SlidingWindow(path) as slide_window:
-            arr_good = slide_window.regression(1, 2, 5)
-            arr_brute = rbg.regression_brute(1, 2, 5)
+        band_num_1 = 1
+        band_num_2 = 2
+        try:
+            with SlidingWindow(self.random) as slide_window:
+                slide_window.convert_image = False
+                slide_window.tif_dtype = self.dtype
+                regression_path = slide_window.regression(band_num_1, band_num_2, self.num_aggre)
+                with rasterio.open(regression_path) as img:
+                    arr_good = img.read(1)
+
+                band_1 = slide_window._img.read(band_num_1).astype(self.dtype)
+                band_2 = slide_window._img.read(band_num_2).astype(self.dtype)
+                
+            arr_brute = rbg.regression_brute(band_1, band_2, self.num_aggre)
+            
+        finally:
+            if (os.path.exists(regression_path)):
+                os.remove(regression_path)
 
         self.assertTrue(np.allclose(arr_good, arr_brute))
 
-    def test_dem_aggregation(self):
-        path = self.img_gen.random()
-        with SlidingWindow(path) as slide_window:
-            slide_window.initialize_dem()
-            slide_window.aggregate_dem(5)
-            dem_data = slide_window._dem_data
-
-            slide_window.initialize_dem()
-            slide_window._aggregate_dem_brute(5)
-            dem_data_brute = slide_window._dem_data
-
-        self.assertTrue(np.array_equal(dem_data.z(), dem_data_brute.z()))
-        self.assertTrue(np.array_equal(dem_data.xz(), dem_data_brute.xz()))
-        self.assertTrue(np.array_equal(dem_data.yz(), dem_data_brute.yz()))
-        self.assertTrue(np.array_equal(dem_data.xxz(), dem_data_brute.xxz()))
-        self.assertTrue(np.array_equal(dem_data.yyz(), dem_data_brute.yyz()))
-        self.assertTrue(np.array_equal(dem_data.xyz(), dem_data_brute.xyz()))
-
     def test_data_init(self):
-        path = self.img_gen.random(image_size=self.IMAGE_SIZE, num_bands=1)
-        with rasterio.open(path) as img:
-            arr = img.read(1).astype(float)
-        with SlidingWindow(path) as slide_window:
+        with SlidingWindow(self.random) as slide_window:
+            arr = slide_window._img.read(1).astype(self.dtype)
             slide_window.initialize_dem()
             arr_dict = slide_window._dem_data
             arr_zeros = np.zeros_like(arr)
@@ -130,135 +130,89 @@ class TestSlidingWindow(unittest.TestCase):
 
 
     def test_data_aggregation(self):
-        path = self.img_gen.random(image_size=self.IMAGE_SIZE, num_bands=1)
-        with rasterio.open(path) as img:
-            arr = img.read(1).astype(float)
-        with SlidingWindow(path) as slide_window:
+        with SlidingWindow(self.random) as slide_window:
+            arr = slide_window._img.read(1).astype(self.dtype)
             slide_window.initialize_dem()
-            slide_window.aggregate_dem(self.NUM_AGGRE)
+            slide_window.aggregate_dem(self.num_aggre)
             arr_dict = slide_window._dem_data
             
-            self.assertTrue(np.array_equal(arr_dict.z(), aggregation.aggregate_z_brute(arr, self.NUM_AGGRE)))
-            self.assertTrue(np.array_equal(arr_dict.xz(), aggregation.aggregate_xz_brute(arr, self.NUM_AGGRE)))
-            self.assertTrue(np.array_equal(arr_dict.yz(), aggregation.aggregate_yz_brute(arr, self.NUM_AGGRE)))
-            self.assertTrue(np.array_equal(arr_dict.xxz(), aggregation.aggregate_xxz_brute(arr, self.NUM_AGGRE)))
-            self.assertTrue(np.array_equal(arr_dict.yyz(), aggregation.aggregate_yyz_brute(arr, self.NUM_AGGRE))) 
-            self.assertTrue(np.array_equal(arr_dict.xyz(), aggregation.aggregate_xyz_brute(arr, self.NUM_AGGRE)))
-
-    def test_create_tif_transform(self):
-        agg_num_1 = 2
-        agg_num_2 = 5
-
-        pix_agg_1 = 2**agg_num_1
-        pix_agg_2 = 2**agg_num_2
-        paths = [self.img_gen.random(), self.img_gen.random(angle=45), self.img_gen.random(angle=75)]
-        # initalize to remove files later
-        path1 = path2 = ''
-        for i in range(len(paths)):
-            try:
-                with self.subTest(path=paths[i]):
-                    with SlidingWindow(paths[i]) as slide_window:
-                        slide_window.dem_initialize_arrays()
-                        slide_window.dem_aggregation_step(agg_num_1)
-                        path1 = slide_window.dem_mean('z')
-                        slide_window.dem_aggregation_step(agg_num_2 - agg_num_1)
-                        path2 = slide_window.dem_mean('z')
-
-                    with rasterio.open(paths[i]) as img:
-                        transform = img.profile['transform']
-                    with rasterio.open(path1) as img:
-                        transform_1 = img.profile['transform']
-                    with rasterio.open(path2) as img:
-                        transform_2 = img.profile['transform']
-
-                    map_width = math.sqrt(transform[0]**2 + transform[3]**2)
-                    map_height = math.sqrt(transform[1]**2 + transform[4]**2)
-                    self.assertTrue(transform_1[0] == transform[0]*pix_agg_1)
-                    self.assertTrue(transform_1[1] == transform[1]*pix_agg_1)
-                    self.assertTrue(transform_1[2] == (transform[2] + (map_width*(pix_agg_1 - 1)/2)))
-                    self.assertTrue(transform_1[3] == transform[3]*pix_agg_1)
-                    self.assertTrue(transform_1[4] == transform[4]*pix_agg_1)
-                    self.assertTrue(transform_1[5] == (transform[5] - (map_height*(pix_agg_1 - 1)/2)))
-
-                    self.assertTrue(transform_2[0] == transform[0]*pix_agg_2)
-                    self.assertTrue(transform_2[1] == transform[1]*pix_agg_2)
-                    self.assertTrue(transform_2[2] == (transform[2] + (map_width*(pix_agg_2 - 1)/2)))
-                    self.assertTrue(transform_2[3] == transform[3]*pix_agg_2)
-                    self.assertTrue(transform_2[4] == transform[4]*pix_agg_2)
-                    self.assertTrue(transform_2[5] == (transform[5] - (map_height*(pix_agg_2 - 1)/2)))
-            finally:
-                if (os.path.exists(path1)):
-                    os.remove(path1)
-                if (os.path.exists(path2)):
-                        os.remove(path2)
-
-    def test_create_tif_export(self):
-        num_aggre = 5
-        path = self.img_gen.random()
-        try:
-            with SlidingWindow(path) as slide_window:
-                slide_window.dem_initialize_arrays()
-                slide_window.dem_aggregation_step(num_aggre)
-                new_path = slide_window.dem_export_arrays()
-            self.assertTrue(os.path.exists(new_path))
-
-            with rasterio.open(path) as img:
-                transform = img.profile['transform']
-            with rasterio.open(new_path) as img:
-                transform_new = img.profile['transform']
-
-            with SlidingWindow(new_path) as slide_window:
-                slide_window.dem_import_arrays()
-                self.assertTrue(slide_window.dem_pixels_aggre == 2**num_aggre)
-                self.assertTrue(len(slide_window.dem_arr_dict) == 6)
-
-            for i in range(6):
-                self.assertTrue(transform_new[i] == transform[i])
-        finally:
-            if (os.path.exists(new_path)):
-                os.remove(new_path)
+            self.assertTrue(np.array_equal(arr_dict.z(), aggregation.aggregate_z_brute(arr, self.num_aggre)))
+            self.assertTrue(np.array_equal(arr_dict.xz(), aggregation.aggregate_xz_brute(arr, self.num_aggre)))
+            self.assertTrue(np.array_equal(arr_dict.yz(), aggregation.aggregate_yz_brute(arr, self.num_aggre)))
+            self.assertTrue(np.array_equal(arr_dict.xxz(), aggregation.aggregate_xxz_brute(arr, self.num_aggre)))
+            self.assertTrue(np.array_equal(arr_dict.yyz(), aggregation.aggregate_yyz_brute(arr, self.num_aggre))) 
+            self.assertTrue(np.array_equal(arr_dict.xyz(), aggregation.aggregate_xyz_brute(arr, self.num_aggre)))
 
     def test_aspect(self):
-        path = self.img_gen.se_gradient()
-        with SlidingWindow(path) as slide_window:
-            slide_window.dem_initialize_arrays()
-            slide_window.dem_aggregation_step(5)
-            aspect_path = ''
+        with SlidingWindow(self.gauss) as slide_window:
+            slide_window.convert_image = False
+            slide_window.initialize_dem()
+            slide_window.aggregate_dem(self.num_aggre)
             try:
                 aspect_path = slide_window.dem_aspect()
-                slope_angle_path = slide_window.dem_slope_angle()
                 with rasterio.open(aspect_path) as img:
-                    arr_aspect = img.read(1).astype(float)
-                with rasterio.open(slope_angle_path) as img:
-                    arr_slope_angle = img.read(1).astype(float)
+                    arr_aspect = img.read(1)
 
-                aspect_angle_perc = (7*math.pi/4)/(2*math.pi)
-                aspect_value = aspect_angle_perc*np.iinfo(self.img_gen.dtype).max
-                self.assertTrue(np.all(arr_aspect == math.floor(aspect_value)))
+                # assumed to be square
+                size = arr_aspect.shape[0]
+                for y in range(size):
+                    for x in range(size):
+                        aspect = self.img_gen.gauss_aspect_point(x, y, size)
+                        self.assertTrue(
+                            math.isclose(arr_aspect[y, x] + (2 * math.pi), aspect + (2 * math.pi), rel_tol=1e-4)
+                            or
+                            math.isclose(arr_aspect[y, x] + (2 * math.pi), aspect, rel_tol=1e-4)
+                            or
+                            math.isclose(arr_aspect[y, x], aspect + (2 * math.pi), rel_tol=1e-4)
+                        )
 
-                slope_angle_perc = math.atan(2/math.sqrt(2))/(math.pi/2)
-                slope_value = slope_angle_perc*np.iinfo(self.img_gen.dtype).max
-                self.assertTrue(np.all(arr_slope_angle == math.floor(slope_value)))
             finally:
                 if (os.path.exists(aspect_path)):
                     os.remove(aspect_path)
-                if (os.path.exists(slope_angle_path)):
-                    os.remove(slope_angle_path)
 
-    def test_profile(self):
-        image_size = 300
-        sigma = image_size/4
-        path = self.img_gen.gauss()
-        with SlidingWindow(path) as slide_window:
-            slide_window.dem_initialize_arrays()
-            slide_window.dem_aggregation_step(5)
-            arr_profile = slide_window._profile()
-            for y in range(arr_profile.shape[0]):
-                for x in range(arr_profile.shape[1]):
-                    test_val = 2*math.pi*math.sqrt( (math.exp((2*(x**2+y**2))/sigma**2)*sigma**8*(-sigma**2+x**2+y**2)**2) / (2*math.exp((x**2+y**2)/sigma**2)*math.pi*sigma**6+x**2+y**2)**3 )
-                    my_value = arr_profile[y][x]
-                    math.isclose(my_value, test_val)
-                
+    def test_standard(self):
+        with SlidingWindow(self.gauss) as slide_window:
+            slide_window.convert_image = False
+            slide_window.initialize_dem()
+            slide_window.aggregate_dem(self.num_aggre)
+            try:
+                standard_path = slide_window.dem_standard()
+                with rasterio.open(standard_path) as img:
+                    arr_standard = img.read(1)
+
+                # assumed to be square
+                size = arr_standard.shape[0]
+                for y in range(size):
+                    for x in range(size):
+                        standard = self.img_gen.gauss_standard_point(x, y, size)
+
+                math.isclose(arr_standard[y, x], standard)
+
+            finally:
+                if (os.path.exists(standard_path)):
+                    os.remove(standard_path)
+
+    def test_slope(self):
+        with SlidingWindow(self.gauss) as slide_window:
+            slide_window.convert_image = False
+            slide_window.initialize_dem()
+            slide_window.aggregate_dem(self.num_aggre)
+            try:
+                slope_path = slide_window.dem_slope()
+                with rasterio.open(slope_path) as img:
+                    arr_slope = img.read(1)
+
+                # assumed to be square
+                size = arr_slope.shape[0]
+                for y in range(size):
+                    for x in range(size):
+                        slope = self.img_gen.gauss_standard_point(x, y, size)
+
+                math.isclose(arr_slope[y, x], slope)
+
+            finally:
+                if (os.path.exists(slope_path)):
+                    os.remove(slope_path)
 
 if __name__ == '__main__':
     unittest.main()
